Change LB algorithm test based on encountered production issue
