+++ Dockerfile
@@ -101,7 +101,7 @@ FROM builder AS testbuilder
 ARG DPSERVICE_FEATURES=""
 RUN meson setup release_build $DPSERVICE_FEATURES --buildtype=release && ninja -C release_build
 RUN CC=clang CXX=clang++ meson setup clang_build $DPSERVICE_FEATURES && ninja -C clang_build
-RUN meson setup xtratest_build $DPSERVICE_FEATURES -Denable_tests=true -Denable_static_underlay_ip=true && ninja -C xtratest_build
+RUN meson setup xtratest_build $DPSERVICE_FEATURES -Denable_tests=true && ninja -C xtratest_build
 
 
 # Test-image to run pytest
